general fixes and moved mock testing classes to a separate file. once testing is done can simply delete file instead of editing all the classes
